+++ src/collective/cover/tests/test_list_tile.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from collective.cover.controlpanel import ICoverSettings
 from collective.cover.testing import ALL_CONTENT_TYPES
 from collective.cover.tests.base import TestTileMixin
 from collective.cover.tests.utils import today
@@ -9,8 +10,10 @@ from plone import api
 from plone.app.testing import login
 from plone.app.testing import logout
 from plone.app.testing import TEST_USER_NAME
+from plone.registry.interfaces import IRegistry
 from plone.tiles.interfaces import ITileDataManager
 from plone.uuid.interfaces import IUUID
+from zope.component import getUtility
 
 import unittest
 
@@ -260,3 +263,47 @@ class ListTileTestCase(TestTileMixin, unittest.TestCase):
         obj = self.portal['my-image']
         self.assertTrue(self.tile.thumbnail(obj))
         self.assertIsInstance(self.tile(), unicode)
+
+    def test_populate_with_collection(self):
+        from collective.cover.testing import zptlogo
+        from collective.cover.tests.utils import set_image_field
+        with api.env.adopt_roles(['Manager']):
+            api.content.create(self.portal, 'News Item', id='new1')
+            api.content.create(self.portal, 'News Item', id='new2')
+            api.content.create(self.portal, 'News Item', id='new3')
+            # handle Archetypes and Dexterity
+            set_image_field(self.portal['new1'], zptlogo)
+            set_image_field(self.portal['new2'], zptlogo)
+
+            query = [dict(
+                i='portal_type',
+                o='plone.app.querystring.operation.selection.is',
+                v='News Item',
+            )]
+            col = api.content.create(
+                self.portal, 'Collection', 'collection', query=query)
+
+        self.tile.populate_with_object(col)
+        rendered = self.tile()
+        self.assertIn(u'<h2><a href="http://nohost/plone/collection', rendered)
+
+    def test_populate_with_folder(self):
+        # Add folder to default accepted_ct
+        registry = getUtility(IRegistry)
+        settings = registry.forInterface(ICoverSettings)
+        settings.searchable_content_types += ['Folder']
+
+        from collective.cover.testing import zptlogo
+        from collective.cover.tests.utils import set_image_field
+        with api.env.adopt_roles(['Manager']):
+            folder = api.content.create(self.portal, 'Folder', id='folder')
+            api.content.create(folder, 'News Item', id='new1')
+            api.content.create(folder, 'News Item', id='new2')
+            api.content.create(folder, 'News Item', id='new3')
+            # handle Archetypes and Dexterity
+            set_image_field(folder['new1'], zptlogo)
+            set_image_field(folder['new2'], zptlogo)
+
+        self.tile.populate_with_object(folder)
+        rendered = self.tile()
+        self.assertIn(u'<h2><a href="http://nohost/plone/folder', rendered)