+++ src/collective/cover/tiles/carousel.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from collective.cover import _
-from collective.cover.interfaces import ICoverUIDsProvider
 from collective.cover.interfaces import ITileEditForm
 from collective.cover.tiles.list import IListTile
 from collective.cover.tiles.list import ListTile
@@ -9,6 +8,7 @@ from collective.cover.widgets.interfaces import ITextLinesSortableWidget
 from collective.cover.widgets.textlinessortable import TextLinesSortableFieldWidget
 from plone.app.uuid.utils import uuidToObject
 from plone.autoform import directives as form
+from plone.memoize import view
 from plone.tiles.interfaces import ITileDataManager
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from z3c.form.converter import BaseDataConverter
@@ -66,11 +66,20 @@ class CarouselTile(ListTile):
         :type uuids: List of strings
         """
         super(ListTile, self).populate_with_object(obj)  # check permission
-        uuids = ICoverUIDsProvider(obj).getUIDs()
+        if obj.portal_type in ('Topic', 'Collection'):
+            uuids = [i.UID for i in obj.queryCatalog()]
+        else:
+            uuids = [self.get_uuid(obj)]
         # accept just elements with a lead image
         uuids = [i for i in uuids if self._has_image_field(uuidToObject(i))]
         if uuids:
             self.populate_with_uuids(uuids)
+
+    @view.memoize
+    def accepted_ct(self):
+        """Return all content types available (default value)."""
+        cts = set(super(CarouselTile, self).accepted_ct())
+        return list(cts - frozenset(['Folder']))
 
     def autoplay(self):
         if self.data['autoplay'] is None: